+++ lib/packages/contracts/src/index.ts
@@ -1,4 +1,5 @@
-/** @module core */
+/** @module contracts */
 
 export * from './getContractDatablock';
 export * from './ContractHelper';
+export * from './generateContractBytes';