+++ lib/packages/contracts/src/typings/args/index.ts
@@ -1,0 +1,8 @@
+/** @module contracts */
+
+/**
+ * Copyright (c) 2019 Burst Apps Team
+ */
+
+export * from './generateContractBytesArgs';
+