+++ lib/packages/core/src/typings/args/index.ts
@@ -4,4 +4,4 @@ export * from './sendAmountArgs';
 export * from './sendMessageArgs';
 export * from './sendEncryptedMessageArgs';
 export * from './getAccountTransactionsArgs';
-export * from './generateContractBytesArgs';
+export * from '../../../../contracts/src/typings/args/generateContractBytesArgs';